Finished basis for parsing ACPI v1.
Minor Updates
- Modified paging code to have 2 implementations
	- paging_map(virt_addr, flags) --> Removes need to call palloc_physical() and then pass that as an argument
	- paging_map2(phys_addr, virt_addr, flags, mapping_flags) --> Provide more control over where + how a page will be mapped into memory

+++ kernel/includes/kernel/mm/palloc.h
@@ -5,10 +5,10 @@
 
 #define PALLOC_INITIAL_BITMAP_SIZE 256
 
-void palloc_init(uint32_t low_address);
-void palloc_init2(uint32_t low_address, struct multiboot_tag_mmap *mb_mmap);
+void palloc_init(uintptr_t low_address);
+void palloc_init2(uintptr_t low_address, struct multiboot_tag_mmap *mb_mmap);
 
-uint32_t palloc_physical();
-void palloc_release(uint32_t address);
+uintptr_t palloc_physical();
+void palloc_release(uintptr_t address);
 
-void palloc_mark_inuse(uint32_t address);
+void palloc_mark_inuse(uintptr_t address);
